refactor: re-organize login
