+++ Cargo.toml
@@ -35,10 +35,15 @@ all = "warn"
 [workspace.dependencies]
 clap = { version = "4.5.9", features = ["derive"] }
 cliclack = "0.3.4"
+mockito = "1.5.0"
+path-slash = "0.2.1"
+reqwest = { version = "0.12.5", default-features = false }
+temp-env = { version = "0.3.6", features = ["async_closure"] }
+testdir = "0.9.1"
+thiserror = "1.0.63"
 tokio = { version = "1.38.0", features = [
     "io-util",
     "macros",
     "process",
     "rt-multi-thread",
 ] }
-thiserror = "1.0.63"
