+++ crates/commands/src/commands/login.rs
@@ -1,5 +1,11 @@
 use clap::Parser;
-use soldeer_core::{auth::login, Result};
+use cliclack::{input, log::remark};
+use email_address_parser::{EmailAddress, ParsingOptions};
+use soldeer_core::{
+    auth::{execute_login, Credentials},
+    errors::AuthError,
+    Result,
+};
 
 /// Log into the central repository to push packages
 #[derive(Debug, Clone, Default, Parser)]
@@ -8,5 +14,27 @@ pub struct Login {}
 
 pub(crate) async fn login_command() -> Result<()> {
     login().await?;
+    Ok(())
+}
+
+async fn login() -> std::result::Result<(), AuthError> {
+    remark("If you do not have an account, please visit soldeer.xyz to create one.")?;
+
+    let email: String = input("Email address")
+        .validate(|input: &String| {
+            if input.is_empty() {
+                Err("Email is required")
+            } else {
+                match EmailAddress::parse(input, Some(ParsingOptions::default())) {
+                    None => Err("Invalid email address"),
+                    Some(_) => Ok(()),
+                }
+            }
+        })
+        .interact()?;
+
+    let password = cliclack::password("Password").mask('▪').interact()?;
+
+    execute_login(&Credentials { email, password }).await?;
     Ok(())
 }
